# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ERP MIF Maroc - Development",
    "description": "Backend ERP pour la gestion des interventions industrielles",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/techniciens/": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Lister les techniciens",
        "description": "Liste tous les techniciens.",
        "operationId": "list_techniciens_api_v1_techniciens__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TechnicienOut"
                  },
                  "type": "array",
                  "title": "Response List Techniciens Api V1 Techniciens  Get"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "techniciens"
        ],
        "summary": "Créer un technicien",
        "description": "Crée un nouveau technicien (réservé aux responsables).",
        "operationId": "create_new_technicien_api_v1_techniciens__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicienCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/techniciens/competences": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Lister les compétences",
        "description": "Liste toutes les compétences.",
        "operationId": "list_competences_api_v1_techniciens_competences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompetenceOut"
                  },
                  "type": "array",
                  "title": "Response List Competences Api V1 Techniciens Competences Get"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "techniciens"
        ],
        "summary": "Créer une compétence",
        "description": "Crée une nouvelle compétence (réservé aux responsables).",
        "operationId": "create_new_competence_api_v1_techniciens_competences_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetenceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetenceOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/techniciens/{technicien_id}": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Détail d’un technicien",
        "description": "Récupère le détail d’un technicien par ID.",
        "operationId": "get_technicien_api_v1_techniciens__technicien_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "techniciens"
        ],
        "summary": "Mettre à jour un technicien",
        "operationId": "update_technicien_api_v1_techniciens__technicien_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicienBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "techniciens"
        ],
        "summary": "Supprimer un technicien",
        "operationId": "delete_technicien_api_v1_techniciens__technicien_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/techniciens/": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Lister les techniciens",
        "description": "Liste tous les techniciens.",
        "operationId": "list_techniciens_techniciens__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TechnicienOut"
                  },
                  "type": "array",
                  "title": "Response List Techniciens Techniciens  Get"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "techniciens"
        ],
        "summary": "Créer un technicien",
        "description": "Crée un nouveau technicien (réservé aux responsables).",
        "operationId": "create_new_technicien_techniciens__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicienCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/techniciens/competences": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Lister les compétences",
        "description": "Liste toutes les compétences.",
        "operationId": "list_competences_techniciens_competences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompetenceOut"
                  },
                  "type": "array",
                  "title": "Response List Competences Techniciens Competences Get"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "techniciens"
        ],
        "summary": "Créer une compétence",
        "description": "Crée une nouvelle compétence (réservé aux responsables).",
        "operationId": "create_new_competence_techniciens_competences_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompetenceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetenceOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/techniciens/{technicien_id}": {
      "get": {
        "tags": [
          "techniciens"
        ],
        "summary": "Détail d’un technicien",
        "description": "Récupère le détail d’un technicien par ID.",
        "operationId": "get_technicien_techniciens__technicien_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "techniciens"
        ],
        "summary": "Mettre à jour un technicien",
        "operationId": "update_technicien_techniciens__technicien_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechnicienBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicienOut"
                }
              }
            }
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "techniciens"
        ],
        "summary": "Supprimer un technicien",
        "operationId": "delete_technicien_techniciens__technicien_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "technicien_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Technicien Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Technicien ou compétence introuvable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompetenceCreate": {
        "properties": {
          "nom": {
            "type": "string",
            "title": "Nom"
          }
        },
        "type": "object",
        "required": [
          "nom"
        ],
        "title": "CompetenceCreate",
        "description": "Données nécessaires à la création d'une compétence."
      },
      "CompetenceOut": {
        "properties": {
          "nom": {
            "type": "string",
            "title": "Nom"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "nom",
          "id"
        ],
        "title": "CompetenceOut",
        "description": "Schéma utilisé en sortie API (inclut l'id)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TechnicienBase": {
        "properties": {
          "equipe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipe"
          },
          "disponibilite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disponibilite"
          }
        },
        "type": "object",
        "title": "TechnicienBase",
        "description": "Champs communs à tous les techniciens."
      },
      "TechnicienCreate": {
        "properties": {
          "equipe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipe"
          },
          "disponibilite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disponibilite"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "competences_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competences Ids"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "TechnicienCreate",
        "description": "Payload attendu pour créer un technicien."
      },
      "TechnicienOut": {
        "properties": {
          "equipe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipe"
          },
          "disponibilite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disponibilite"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          },
          "competences": {
            "items": {
              "$ref": "#/components/schemas/CompetenceOut"
            },
            "type": "array",
            "title": "Competences",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user"
        ],
        "title": "TechnicienOut",
        "description": "Données retournées pour un technicien (lecture/detail)."
      },
...
      "UserOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "username",
          "full_name",
          "email",
          "role",
          "id",
          "is_active"
        ],
        "title": "UserOut",
        "description": "Données renvoyées par l'API :\n- toutes les infos utilisateur sauf le mot de passe\n- audit : id, statut, timestamps"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "responsable",
          "technicien",
          "client"
        ],
        "title": "UserRole",
        "description": "Rôles disponibles dans l’ERP :\n- admin : contrôle total\n- responsable : supervise les interventions\n- technicien : effectue les interventions\n- client : consultation uniquement"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}